+++ packages/netsuite-adapter/src/constants.ts
@@ -56,6 +56,8 @@ export const PERMITTED_ROLE = 'permittedrole'
 export const RECORD_TYPE = 'recordType'
 export const LAST_FETCH_TIME = '_lastfetchtime'
 export const APPLICATION_ID = 'application_id'
+export const SENDER = 'sender'
+export const RECIPIENTEMAIL = 'recipientemail'
 
 // Field Annotations
 export const IS_ATTRIBUTE = 'isAttribute'
