+++ packages/netsuite-adapter/test/change_validators/account_specific_values.test.ts
@@ -81,75 +81,3 @@ describe('account specific value validator', () => {
     expect(changeErrors[0].elemID).toEqual(instance.elemID)
   })
 })
-
-describe('account specific values validator for sender and recepient fields', () => {
-  let instance: InstanceElement
-  beforeEach(() => {
-    const origInstance = new InstanceElement(
-      'instance',
-      workflowType().type,
-      {
-        isinactive: false,
-        [SCRIPT_ID]: 'customworkflow3',
-        name: 'WokrflowName',
-        workflowstates: {
-          workflowstate: {
-            workflowstate1: {
-              sendmailaction: {
-                workfloraction: {
-                  recipientemail: '[STDUSERUSER]',
-                  recipienttype: 'FIELD',
-                  sender: '[STDUSERUSER]',
-                  sendertype: 'FIELD',
-                },
-              },
-            },
-          },
-        },
-      }
-    )
-    instance = origInstance.clone()
-  })
-
-  it('should not have changeError when deploying an instance without ACCOUNT_SPECIFIC_VALUES', async () => {
-    const after = instance.clone()
-    const changeErrors = await accountSpecificValueValidator(
-      [toChange({ before: instance, after })]
-    )
-    expect(changeErrors).toHaveLength(0)
-  })
-
-  it('should have changeError when deploying an instance with sender = ACCOUNT_SPECIFIC_VALUES', async () => {
-    const after = instance.clone()
-    after.value.sender = '[ACCOUNT_SPECIFIC_VALUE]'
-    const changeErrors = await accountSpecificValueValidator(
-      [toChange({ before: instance, after })]
-    )
-    expect(changeErrors).toHaveLength(1)
-    expect(changeErrors[0].severity).toEqual('Error')
-    expect(changeErrors[0].elemID).toEqual(instance.elemID)
-  })
-
-  it('should have changeError when deploying an instance with recepientemail = ACCOUNT_SPECIFIC_VALUES', async () => {
-    const after = instance.clone()
-    after.value.recipientemail = '[ACCOUNT_SPECIFIC_VALUE]'
-    const changeErrors = await accountSpecificValueValidator(
-      [toChange({ before: instance, after })]
-    )
-    expect(changeErrors).toHaveLength(1)
-    expect(changeErrors[0].severity).toEqual('Error')
-    expect(changeErrors[0].elemID).toEqual(instance.elemID)
-  })
-
-  it('should have changeError when deploying an instance with recepientemail = ACCOUNT_SPECIFIC_VALUES and sender = ACCOUNT_SPECIFIC_VALUES', async () => {
-    const after = instance.clone()
-    after.value.recipientemail = '[ACCOUNT_SPECIFIC_VALUE]'
-    after.value.sender = '[ACCOUNT_SPECIFIC_VALUE]'
-    const changeErrors = await accountSpecificValueValidator(
-      [toChange({ before: instance, after })]
-    )
-    expect(changeErrors).toHaveLength(1)
-    expect(changeErrors[0].severity).toEqual('Error')
-    expect(changeErrors[0].elemID).toEqual(instance.elemID)
-  })
-})
